# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.errorResponse:
    properties:
      error:
        properties:
          code:
            example: INVALID_INPUT
            type: string
          message:
            example: validation failed
            type: string
        type: object
    type: object
...
  http.loginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        type: string
    type: object
  http.registerRequest:
    properties:
      email:
        example: user@example.com
        type: string
      inviteCode:
        example: ABC123DEF456
        type: string
      password:
        example: securepassword123
        type: string
    type: object
...
  inbound.UserDTO:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing Ludo tournaments and leagues
  title: Ludo Tournament Management API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user with email and password, returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidates the current session (client should discard token)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inbound.UserDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email, password, and invite code
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"